requirements.txt
beautifulsoup4
cerebras-cloud-sdk
edge-tts
emoji
Flask
groq
httpx
openai
Pillow
pytz
rdkit
replicate
telegraph
python-telegram-bot
chromadb
tiktoken
bytez
